Add in board data structure
This module should build randomized playing fields that are mirrored
over the diagonal so as to keep it fair.  Only supports 2 players at
a time -- seemed to tricky to keep it fair with more than that.

If you want to test it just run

  python board.py

and it should build a board and save a little image representation of
it in out.png

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>